src/third_party/blink/web_tests/http/tests/css/missing-repaint-after-slow-style-sheet-expected.txt

CONSOLE MESSAGE: line 1: Inline script done
CONSOLE MESSAGE: line 1: requestAnimationFrame ran
CONSOLE MESSAGE: line 1: Stylesheet loaded
